user reservations scrollable + expired reservation

+++ components/user/reservation-list/item/item.vue
@@ -1,7 +1,7 @@
 <script setup lang="ts">
 import type { Seat } from "../reservation-list.vue";
 
-const { seat, rideId } = defineProps<{
+const { seat, rideId, expired } = defineProps<{
 	seat: Seat;
 	rideId: string;
 	expired: boolean;
@@ -16,6 +16,8 @@ const emit = defineEmits<{
 const { apiFetch } = useApiFetch();
 
 const handleCancelReservation = async () => {
+	if (expired) return;
+
 	try {
 		await apiFetch("/users/cancelReservations", {
 			method: "DELETE",
